:truck: (circular queue): Move test class to separate header
