+++ libs/ContainerKit/tests/CircularQueue_test.cpp
@@ -2,28 +2,7 @@
 // Copyright 2021 APF France handicap (based on work by Mbed-OS)
 // SPDX-License-Identifier: Apache-2.0
 
-#include "CircularQueue.h"
-#include <array>
-#include <memory>
-
-#include "gtest/gtest.h"
-#include "stubs/mbed/mbed_critical.h"
-
-using namespace leka;
-
-class CircularQueueTest : public testing::Test
-{
-  protected:
-	void SetUp() override
-	{
-		spy_mbed_critical_enter_calls = 0;
-		spy_mbed_critical_exit_calls  = 0;
-	}
-	// void TearDown() override {}
-
-	static constexpr auto TEST_BUFFER_SIZE {10};
-	CircularQueue<int, TEST_BUFFER_SIZE> buf {};
-};
+#include "./CircularQueue_test.h"
 
 TEST_F(CircularQueueTest, initialization)
 {
